+++ Editor/TestReportGraph/SampleGroupAdditionalData.cs
@@ -1,8 +1,11 @@
-﻿struct SampleGroupAdditionalData
+﻿namespace Unity.PerformanceTesting.Editor
 {
-    public float min;
-    public float lowerQuartile;
-    public float median;
-    public float upperQuartile;
-    public float max;
+    struct SampleGroupAdditionalData
+    {
+        public float min;
+        public float lowerQuartile;
+        public float median;
+        public float upperQuartile;
+        public float max;
+    }
 }