+++ Runtime/SampleGroup.cs
@@ -1,13 +1,37 @@
+using System;
+using System.Collections.Generic;
+using Unity.PerformanceTesting.Exceptions;
 using UnityEngine.Profiling;
 
 namespace Unity.PerformanceTesting
 {
-    public class SampleGroup : Unity.PerformanceTesting.Data.SampleGroup
+    [Serializable]
+    public class SampleGroup
     {
-        internal Recorder Recorder;
+        public string Name;
+        public SampleUnit Unit;
+        public bool IncreaseIsBetter;
+        public List<double> Samples = new List<double>();
+        public double Min;
+        public double Max;
+        public double Median;
+        public double Average;
+        public double StandardDeviation;
+        public double Sum;
 
         public SampleGroup(string name, SampleUnit unit = SampleUnit.Millisecond, bool increaseIsBetter = false)
-            : base(name, unit, increaseIsBetter) { }
+        {
+            Name = name;
+            Unit = unit;
+            IncreaseIsBetter = increaseIsBetter;
+
+            if (string.IsNullOrEmpty(name))
+            {
+                throw new PerformanceTestException("Sample group name is empty. Please assign a valid name.");
+            }
+        }
+        
+        internal Recorder Recorder;
 
         public Recorder GetRecorder()
         {